seo implemented, preview added to post details.
