+++ src/OnkelMato.BlogEngine.Core/Service/ISeoProvider.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Mvc.Rendering;
+
+namespace OnkelMato.BlogEngine.Core.Service;
+
+public interface ISeoProvider
+{
+    void SetSiteInfo(IHtmlHelper html, string siteTitle, string siteDescription, string robots = "index, follow");
+}
