+++ src/OnkelMato.BlogEngine.Web/Program.cs
@@ -1,9 +1,6 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Options;
-using OnkelMato.BlogEngine.Core.Configuration;
-using System;
 
 namespace OnkelMato.BlogEngine.Web;
 
@@ -58,41 +55,5 @@ public class Program
         app.UseRouting();
         app.MapRazorPages();
         app.Run();
-    }
-}
-
-// static class for seo extensions
-public static class SeoExtensions
-{
-    public static IServiceCollection AddBlogSeoTags(this IServiceCollection services)
-    {
-        //return services;
-        // todo this must be done later. blog id is unknown here. 
-       
-        //Register your services
-        services.AddSeoTags(seoInfo =>
-        {
-            var serviceProvider = services.BuildServiceProvider();
-            //var db = serviceProvider.GetService<BlogEngineReadRepository>() ?? throw new Exception("Cannot init Database");
-            var blogSettings = serviceProvider.GetService<IOptionsMonitor<BlogConfiguration>>() ?? throw new Exception("Cannot get blog settings");
-            //var blogTitle = db.Blog()?.Title ?? "Onkel Mato Blog Engine";
-            var blogTitle = "Onkel Mato Blog Engine";
-
-            seoInfo.SetSiteInfo(
-                siteTitle: blogTitle,
-                //openSearchUrl: "https://site.com/open-search.xml",  //Optional
-                robots: "index, follow"                             //Optional
-            );
-
-            //Optional
-            //seoInfo.AddFeed(
-            //    title: "Post Feeds",
-            //    url: "https://site.com/rss/",
-            //    feedType: FeedType.Rss);
-
-            //Optional
-            seoInfo.SetLocales(blogSettings.CurrentValue.Language);
-        });
-        return services;
     }
 }
