+++ src/OnkelMato.BlogEngine.Web/appsettings.Development.json
@@ -37,7 +37,7 @@
   },
   "ImportExport": {
     // certificate options
-    //"ValidateCertificates": false,
+    "ValidateCertificates": false,
 
     // JWT token import and export ()
     "EnableJwtImport": true,
@@ -53,8 +53,12 @@
     "EnableJsonExport": true,
     "EnableRemoteExport": true,
 
+    // Sync to a remote page (export). Uses automated JWT upload
+    "EnableSyncToRemote": false,
+
     // Enable Sync from (public) blog engine
-    "EnableBlogSync": true,
+    "EnableBlogExportSync": true,
+    "EnableBlogImportSync": true,
     "BlogSyncSecret": "8E94D58E62FCF811A245BCDFB2484"
   }
 }
