+++ src/OnkelMato.BlogEngine.Web/appsettings.json
@@ -54,7 +54,7 @@
     "EnableJsonExport": false,
 
     // Sync to a remote page (export). Uses automated JWT upload
-    "EnableSyncToRemote" :  false,
+    "EnableSyncToRemote": false,
 
     // Enable Sync from (public) blog engine. Needs to be activated at both ends
     "EnableBlogExportSync": false,
