+++ src/OnkelMato.BlogEngine.Web/wwwroot/css/site.css
@@ -65,3 +65,7 @@ footer {
     background-repeat: no-repeat;
     background-position: center center;
 }
+
+.post-preview {
+    font-style: italic;
+}
