+++ src/OnkelMato.BlogEngine/Pages/PostAdmin/Delete.cshtml
@@ -1,10 +1,11 @@
 ﻿@page
+@using OnkelMato.BlogEngine.Core.Service
 @model OnkelMato.BlogEngine.Pages.PostAdmin.DeleteModel
-
+@inject ISeoProvider SeoProvider
 @{
-    ViewData["Title"] = "Delete";
-    Layout = "~/Pages/Shared/_Layout.cshtml";
+    SeoProvider.SetSiteInfo(Html, "Delete Post", "Blog Engine Post Deletion Page", "noindex, nofollow");
 }
+
 
 <div class="col-md-12">
     <form method="post">
